+++ baseframe/views.py
@@ -89,7 +89,6 @@ def baseframe_context():
     return {
         'networkbar_links': networkbar_links,
         'csrf_form': Form,
-        'ext_assets': ext_assets,
     }