+++ CHANGES.rst
@@ -9,6 +9,9 @@
 * Revamped forms.html and autoform generator
 * ValidEmailDomain validator is now ValidEmail and does syntax checks
 * New NoObfuscatedEmail validator looks for obfuscated emails
+* New FormGenerator constructs forms from field definitions
+* New CoordinatesField renders a map marker drag widget, falling back to lat/lon text fields
+* Codemirror is now an external dependency installed via Makefile
 
 
 0.3.0